feat: Добавлена поддержка дробных номеров в серии
Closes #121

+++ backend.v2/backend.v2/Models/Book.cs
@@ -86,7 +86,7 @@ namespace backend.v2.Models
         /// <summary>
         /// Номер книги в коллекции.
         /// </summary>
-        public short? CollectionOrder {get; set;}
+        public float? CollectionOrder {get; set;}
         
         /// <summary>
         /// Название жанра.
